Refactoring of the QRCode filter to make it resizable and simpler to position.
Refactor the qr code rendering code to make it positionable directly
in the template (wrap the tag in a svg group with a translate
transform attibute).  The parameters now indicate the size of the code
and not its position.

The resizing can also be made in the parent svg group element of the
template, but I find it simpler to give the template filter a goal size
and let it stretch the qrcode to fit it.

This was achieved by using a custom factory to render the svg, which
allowed us to remove the StringIO kludge, remove the namespace and add
a method to resize.

Updated test to reflech qrcode filter changes.