Prepared the flyer svg file to use a new tag for the speaker and title.
The text containing the date line and the speaker/title are split
between two texts.  This allows to set the font definition on the text
instead of in each tspan. Other attributes not directly relevant to
the text positioning were removed.

In order to simplify the coordinates of the tspans, the text is placed
in a translated group.

Note that editing and saving the file using inskape escapes the quotes
used to pass arguments to the filters.

Warning: the exact line height should be 115%, but inkscape does not
accept it, storing it in a separate attribute.